Crate verdant_queries

verdant-queries

Collection of tree-sitter queries for verdant.

See the project overview for more information.

This crate defines constants for three types of tree-sitter queries for lots of parsers. It is intended to be used via verdant-parsers, verdant-parsers-git, or verdant-parsers-gitdep.

The three types of queries are:

highlights: defining the highlight captures for nodes

injections: defining where other languages are injected for highlighting

locals: keeping track of scopes, variables, parameters, etc. to have occurrences of those be highlighted the same everywhere

The constants are defined as <language_name>_<kind>[_CRATES_IO] where <kind> is one of HIGHLIGHTS, INJECTIONS, or LOCALS. The INJECTIONS and LOCALS may be empty for some languages. The constants with the _CRATES_IO suffix aim to be compatible with the latest version of the parser that was published on crates.io. These are used by verdant-parsers, whereas the normal queries are used by verdant-parsers-git and verdant-parsers-gitdep.

The crate source is automatically generated with cargo xtask codegen inside the verdant workspace.

Constants

ADA_HIGHLIGHTS

ADA_HIGHLIGHTS_CRATES_IO

ADA_INJECTIONS

ADA_INJECTIONS_CRATES_IO

ADA_LOCALS

ADA_LOCALS_CRATES_IO

AGDA_HIGHLIGHTS

AGDA_HIGHLIGHTS_CRATES_IO

AGDA_INJECTIONS

AGDA_INJECTIONS_CRATES_IO

AGDA_LOCALS

AGDA_LOCALS_CRATES_IO

APEX_HIGHLIGHTS

APEX_HIGHLIGHTS_CRATES_IO

APEX_INJECTIONS

APEX_INJECTIONS_CRATES_IO

APEX_LOCALS

APEX_LOCALS_CRATES_IO

ARDUINO_HIGHLIGHTS

ARDUINO_HIGHLIGHTS_CRATES_IO

ARDUINO_INJECTIONS

ARDUINO_INJECTIONS_CRATES_IO

ARDUINO_LOCALS

ARDUINO_LOCALS_CRATES_IO

ASM68K_HIGHLIGHTS

ASM68K_HIGHLIGHTS_CRATES_IO

ASM68K_INJECTIONS

ASM68K_INJECTIONS_CRATES_IO

ASM68K_LOCALS

ASM68K_LOCALS_CRATES_IO

ASM_HIGHLIGHTS

ASM_HIGHLIGHTS_CRATES_IO

ASM_INJECTIONS

ASM_INJECTIONS_CRATES_IO

ASM_LOCALS

ASM_LOCALS_CRATES_IO

ASTRO_HIGHLIGHTS

ASTRO_HIGHLIGHTS_CRATES_IO

ASTRO_INJECTIONS

ASTRO_INJECTIONS_CRATES_IO

ASTRO_LOCALS

ASTRO_LOCALS_CRATES_IO

AUTHZED_HIGHLIGHTS

AUTHZED_HIGHLIGHTS_CRATES_IO

AUTHZED_INJECTIONS

AUTHZED_INJECTIONS_CRATES_IO

AUTHZED_LOCALS

AUTHZED_LOCALS_CRATES_IO

AWK_HIGHLIGHTS

AWK_HIGHLIGHTS_CRATES_IO

AWK_INJECTIONS

AWK_INJECTIONS_CRATES_IO

AWK_LOCALS

AWK_LOCALS_CRATES_IO

BASH_HIGHLIGHTS

BASH_HIGHLIGHTS_CRATES_IO

BASH_INJECTIONS

BASH_INJECTIONS_CRATES_IO

BASH_LOCALS

BASH_LOCALS_CRATES_IO

BASS_HIGHLIGHTS

BASS_HIGHLIGHTS_CRATES_IO

BASS_INJECTIONS

BASS_INJECTIONS_CRATES_IO

BASS_LOCALS

BASS_LOCALS_CRATES_IO

BEANCOUNT_HIGHLIGHTS

BEANCOUNT_HIGHLIGHTS_CRATES_IO

BEANCOUNT_INJECTIONS

BEANCOUNT_INJECTIONS_CRATES_IO

BEANCOUNT_LOCALS

BEANCOUNT_LOCALS_CRATES_IO

BIBTEX_HIGHLIGHTS

BIBTEX_HIGHLIGHTS_CRATES_IO

BIBTEX_INJECTIONS

BIBTEX_INJECTIONS_CRATES_IO

BIBTEX_LOCALS

BIBTEX_LOCALS_CRATES_IO

BIB_HIGHLIGHTS

BIB_HIGHLIGHTS_CRATES_IO

BIB_INJECTIONS

BIB_INJECTIONS_CRATES_IO

BIB_LOCALS

BIB_LOCALS_CRATES_IO

BICEP_HIGHLIGHTS

BICEP_HIGHLIGHTS_CRATES_IO

BICEP_INJECTIONS

BICEP_INJECTIONS_CRATES_IO

BICEP_LOCALS

BICEP_LOCALS_CRATES_IO

BITBAKE_HIGHLIGHTS

BITBAKE_HIGHLIGHTS_CRATES_IO

BITBAKE_INJECTIONS

BITBAKE_INJECTIONS_CRATES_IO

BITBAKE_LOCALS

BITBAKE_LOCALS_CRATES_IO

BLADE_HIGHLIGHTS

BLADE_HIGHLIGHTS_CRATES_IO

BLADE_INJECTIONS

BLADE_INJECTIONS_CRATES_IO

BLADE_LOCALS

BLADE_LOCALS_CRATES_IO

BLUEPRINT_HIGHLIGHTS

BLUEPRINT_HIGHLIGHTS_CRATES_IO

BLUEPRINT_INJECTIONS

BLUEPRINT_INJECTIONS_CRATES_IO

BLUEPRINT_LOCALS

BLUEPRINT_LOCALS_CRATES_IO

BPFTRACE_HIGHLIGHTS

BPFTRACE_HIGHLIGHTS_CRATES_IO

BPFTRACE_INJECTIONS

BPFTRACE_INJECTIONS_CRATES_IO

BPFTRACE_LOCALS

BPFTRACE_LOCALS_CRATES_IO

BP_HIGHLIGHTS

BP_HIGHLIGHTS_CRATES_IO

BP_INJECTIONS

BP_INJECTIONS_CRATES_IO

BP_LOCALS

BP_LOCALS_CRATES_IO

BRIGHTSCRIPT_HIGHLIGHTS

BRIGHTSCRIPT_HIGHLIGHTS_CRATES_IO

BRIGHTSCRIPT_INJECTIONS

BRIGHTSCRIPT_INJECTIONS_CRATES_IO

BRIGHTSCRIPT_LOCALS

BRIGHTSCRIPT_LOCALS_CRATES_IO

BZL_HIGHLIGHTS

BZL_HIGHLIGHTS_CRATES_IO

BZL_INJECTIONS

BZL_INJECTIONS_CRATES_IO

BZL_LOCALS

BZL_LOCALS_CRATES_IO

C3_HIGHLIGHTS

C3_HIGHLIGHTS_CRATES_IO

C3_INJECTIONS

C3_INJECTIONS_CRATES_IO

C3_LOCALS

C3_LOCALS_CRATES_IO

CADDY_HIGHLIGHTS

CADDY_HIGHLIGHTS_CRATES_IO

CADDY_INJECTIONS

CADDY_INJECTIONS_CRATES_IO

CADDY_LOCALS

CADDY_LOCALS_CRATES_IO

CAIRO_HIGHLIGHTS

CAIRO_HIGHLIGHTS_CRATES_IO

CAIRO_INJECTIONS

CAIRO_INJECTIONS_CRATES_IO

CAIRO_LOCALS

CAIRO_LOCALS_CRATES_IO

CAPNP_HIGHLIGHTS

CAPNP_HIGHLIGHTS_CRATES_IO

CAPNP_INJECTIONS

CAPNP_INJECTIONS_CRATES_IO

CAPNP_LOCALS

CAPNP_LOCALS_CRATES_IO

CHATITO_HIGHLIGHTS

CHATITO_HIGHLIGHTS_CRATES_IO

CHATITO_INJECTIONS

CHATITO_INJECTIONS_CRATES_IO

CHATITO_LOCALS

CHATITO_LOCALS_CRATES_IO

CIRCOM_HIGHLIGHTS

CIRCOM_HIGHLIGHTS_CRATES_IO

CIRCOM_INJECTIONS

CIRCOM_INJECTIONS_CRATES_IO

CIRCOM_LOCALS

CIRCOM_LOCALS_CRATES_IO

CLARITY_HIGHLIGHTS

CLARITY_HIGHLIGHTS_CRATES_IO

CLARITY_INJECTIONS

CLARITY_INJECTIONS_CRATES_IO

CLARITY_LOCALS

CLARITY_LOCALS_CRATES_IO

CLOJURE_HIGHLIGHTS

CLOJURE_HIGHLIGHTS_CRATES_IO

CLOJURE_INJECTIONS

CLOJURE_INJECTIONS_CRATES_IO

CLOJURE_LOCALS

CLOJURE_LOCALS_CRATES_IO

CMAKE_HIGHLIGHTS

CMAKE_HIGHLIGHTS_CRATES_IO

CMAKE_INJECTIONS

CMAKE_INJECTIONS_CRATES_IO

CMAKE_LOCALS

CMAKE_LOCALS_CRATES_IO

COMMENT_HIGHLIGHTS

COMMENT_HIGHLIGHTS_CRATES_IO

COMMENT_INJECTIONS

COMMENT_INJECTIONS_CRATES_IO

COMMENT_LOCALS

COMMENT_LOCALS_CRATES_IO

COOK_HIGHLIGHTS

COOK_HIGHLIGHTS_CRATES_IO

COOK_INJECTIONS

COOK_INJECTIONS_CRATES_IO

COOK_LOCALS

COOK_LOCALS_CRATES_IO

CORN_HIGHLIGHTS

CORN_HIGHLIGHTS_CRATES_IO

CORN_INJECTIONS

CORN_INJECTIONS_CRATES_IO

CORN_LOCALS

CORN_LOCALS_CRATES_IO

CPON_HIGHLIGHTS

CPON_HIGHLIGHTS_CRATES_IO

CPON_INJECTIONS

CPON_INJECTIONS_CRATES_IO

CPON_LOCALS

CPON_LOCALS_CRATES_IO

CPP_HIGHLIGHTS

CPP_HIGHLIGHTS_CRATES_IO

CPP_INJECTIONS

CPP_INJECTIONS_CRATES_IO

CPP_LOCALS

CPP_LOCALS_CRATES_IO

CSHARP_HIGHLIGHTS

CSHARP_HIGHLIGHTS_CRATES_IO

CSHARP_INJECTIONS

CSHARP_INJECTIONS_CRATES_IO

CSHARP_LOCALS

CSHARP_LOCALS_CRATES_IO

CSS_HIGHLIGHTS

CSS_HIGHLIGHTS_CRATES_IO

CSS_INJECTIONS

CSS_INJECTIONS_CRATES_IO

CSS_LOCALS

CSS_LOCALS_CRATES_IO

CSV_HIGHLIGHTS

CSV_HIGHLIGHTS_CRATES_IO

CSV_INJECTIONS

CSV_INJECTIONS_CRATES_IO

CSV_LOCALS

CSV_LOCALS_CRATES_IO

CUDA_HIGHLIGHTS

CUDA_HIGHLIGHTS_CRATES_IO

CUDA_INJECTIONS

CUDA_INJECTIONS_CRATES_IO

CUDA_LOCALS

CUDA_LOCALS_CRATES_IO

CUE_HIGHLIGHTS

CUE_HIGHLIGHTS_CRATES_IO

CUE_INJECTIONS

CUE_INJECTIONS_CRATES_IO

CUE_LOCALS

CUE_LOCALS_CRATES_IO

CYLC_HIGHLIGHTS

CYLC_HIGHLIGHTS_CRATES_IO

CYLC_INJECTIONS

CYLC_INJECTIONS_CRATES_IO

CYLC_LOCALS

CYLC_LOCALS_CRATES_IO

C_HIGHLIGHTS

C_HIGHLIGHTS_CRATES_IO

C_INJECTIONS

C_INJECTIONS_CRATES_IO

C_LOCALS

C_LOCALS_CRATES_IO

C_SHARP_HIGHLIGHTS

C_SHARP_HIGHLIGHTS_CRATES_IO

C_SHARP_INJECTIONS

C_SHARP_INJECTIONS_CRATES_IO

C_SHARP_LOCALS

C_SHARP_LOCALS_CRATES_IO

DART_HIGHLIGHTS

DART_HIGHLIGHTS_CRATES_IO

DART_INJECTIONS

DART_INJECTIONS_CRATES_IO

DART_LOCALS

DART_LOCALS_CRATES_IO

DESKTOP_HIGHLIGHTS

DESKTOP_HIGHLIGHTS_CRATES_IO

DESKTOP_INJECTIONS

DESKTOP_INJECTIONS_CRATES_IO

DESKTOP_LOCALS

DESKTOP_LOCALS_CRATES_IO

DHALL_HIGHLIGHTS

DHALL_HIGHLIGHTS_CRATES_IO

DHALL_INJECTIONS

DHALL_INJECTIONS_CRATES_IO

DHALL_LOCALS

DHALL_LOCALS_CRATES_IO

DIFF_HIGHLIGHTS

DIFF_HIGHLIGHTS_CRATES_IO

DIFF_INJECTIONS

DIFF_INJECTIONS_CRATES_IO

DIFF_LOCALS

DIFF_LOCALS_CRATES_IO

DISASSEMBLY_HIGHLIGHTS

DISASSEMBLY_HIGHLIGHTS_CRATES_IO

DISASSEMBLY_INJECTIONS

DISASSEMBLY_INJECTIONS_CRATES_IO

DISASSEMBLY_LOCALS

DISASSEMBLY_LOCALS_CRATES_IO

DJOT_HIGHLIGHTS

DJOT_HIGHLIGHTS_CRATES_IO

DJOT_INJECTIONS

DJOT_INJECTIONS_CRATES_IO

DJOT_LOCALS

DJOT_LOCALS_CRATES_IO

DOCKERFILE_HIGHLIGHTS

DOCKERFILE_HIGHLIGHTS_CRATES_IO

DOCKERFILE_INJECTIONS

DOCKERFILE_INJECTIONS_CRATES_IO

DOCKERFILE_LOCALS

DOCKERFILE_LOCALS_CRATES_IO

DOT_HIGHLIGHTS

DOT_HIGHLIGHTS_CRATES_IO

DOT_INJECTIONS

DOT_INJECTIONS_CRATES_IO

DOT_LOCALS

DOT_LOCALS_CRATES_IO

DOXYGEN_HIGHLIGHTS

DOXYGEN_HIGHLIGHTS_CRATES_IO

DOXYGEN_INJECTIONS

DOXYGEN_INJECTIONS_CRATES_IO

DOXYGEN_LOCALS

DOXYGEN_LOCALS_CRATES_IO

DTD_HIGHLIGHTS

DTD_HIGHLIGHTS_CRATES_IO

DTD_INJECTIONS

DTD_INJECTIONS_CRATES_IO

DTD_LOCALS

DTD_LOCALS_CRATES_IO

DTS_HIGHLIGHTS

DTS_HIGHLIGHTS_CRATES_IO

DTS_INJECTIONS

DTS_INJECTIONS_CRATES_IO

DTS_LOCALS

DTS_LOCALS_CRATES_IO

D_HIGHLIGHTS

D_HIGHLIGHTS_CRATES_IO

D_INJECTIONS

D_INJECTIONS_CRATES_IO

D_LOCALS

D_LOCALS_CRATES_IO

EARTHFILE_HIGHLIGHTS

EARTHFILE_HIGHLIGHTS_CRATES_IO

EARTHFILE_INJECTIONS

EARTHFILE_INJECTIONS_CRATES_IO

EARTHFILE_LOCALS

EARTHFILE_LOCALS_CRATES_IO

EBNF_HIGHLIGHTS

EBNF_HIGHLIGHTS_CRATES_IO

EBNF_INJECTIONS

EBNF_INJECTIONS_CRATES_IO

EBNF_LOCALS

EBNF_LOCALS_CRATES_IO

EDITORCONFIG_HIGHLIGHTS

EDITORCONFIG_HIGHLIGHTS_CRATES_IO

EDITORCONFIG_INJECTIONS

EDITORCONFIG_INJECTIONS_CRATES_IO

EDITORCONFIG_LOCALS

EDITORCONFIG_LOCALS_CRATES_IO

EDS_HIGHLIGHTS

EDS_HIGHLIGHTS_CRATES_IO

EDS_INJECTIONS

EDS_INJECTIONS_CRATES_IO

EDS_LOCALS

EDS_LOCALS_CRATES_IO

EELIXIR_HIGHLIGHTS

EELIXIR_HIGHLIGHTS_CRATES_IO

EELIXIR_INJECTIONS

EELIXIR_INJECTIONS_CRATES_IO

EELIXIR_LOCALS

EELIXIR_LOCALS_CRATES_IO

EJS_HIGHLIGHTS

EJS_HIGHLIGHTS_CRATES_IO

EJS_INJECTIONS

EJS_INJECTIONS_CRATES_IO

EJS_LOCALS

EJS_LOCALS_CRATES_IO

ELISP_HIGHLIGHTS

ELISP_HIGHLIGHTS_CRATES_IO

ELISP_INJECTIONS

ELISP_INJECTIONS_CRATES_IO

ELISP_LOCALS

ELISP_LOCALS_CRATES_IO

ELIXIR_HIGHLIGHTS

ELIXIR_HIGHLIGHTS_CRATES_IO

ELIXIR_INJECTIONS

ELIXIR_INJECTIONS_CRATES_IO

ELIXIR_LOCALS

ELIXIR_LOCALS_CRATES_IO

ELM_HIGHLIGHTS

ELM_HIGHLIGHTS_CRATES_IO

ELM_INJECTIONS

ELM_INJECTIONS_CRATES_IO

ELM_LOCALS

ELM_LOCALS_CRATES_IO

ELSA_HIGHLIGHTS

ELSA_HIGHLIGHTS_CRATES_IO

ELSA_INJECTIONS

ELSA_INJECTIONS_CRATES_IO

ELSA_LOCALS

ELSA_LOCALS_CRATES_IO

ELVISH_HIGHLIGHTS

ELVISH_HIGHLIGHTS_CRATES_IO

ELVISH_INJECTIONS

ELVISH_INJECTIONS_CRATES_IO

ELVISH_LOCALS

ELVISH_LOCALS_CRATES_IO

ENFORCE_HIGHLIGHTS

ENFORCE_HIGHLIGHTS_CRATES_IO

ENFORCE_INJECTIONS

ENFORCE_INJECTIONS_CRATES_IO

ENFORCE_LOCALS

ENFORCE_LOCALS_CRATES_IO

ERB_HIGHLIGHTS

ERB_HIGHLIGHTS_CRATES_IO

ERB_INJECTIONS

ERB_INJECTIONS_CRATES_IO

ERB_LOCALS

ERB_LOCALS_CRATES_IO

ERLANG_HIGHLIGHTS

ERLANG_HIGHLIGHTS_CRATES_IO

ERLANG_INJECTIONS

ERLANG_INJECTIONS_CRATES_IO

ERLANG_LOCALS

ERLANG_LOCALS_CRATES_IO

ERUBY_HIGHLIGHTS

ERUBY_HIGHLIGHTS_CRATES_IO

ERUBY_INJECTIONS

ERUBY_INJECTIONS_CRATES_IO

ERUBY_LOCALS

ERUBY_LOCALS_CRATES_IO

FAUST_HIGHLIGHTS

FAUST_HIGHLIGHTS_CRATES_IO

FAUST_INJECTIONS

FAUST_INJECTIONS_CRATES_IO

FAUST_LOCALS

FAUST_LOCALS_CRATES_IO

FENNEL_HIGHLIGHTS

FENNEL_HIGHLIGHTS_CRATES_IO

FENNEL_INJECTIONS

FENNEL_INJECTIONS_CRATES_IO

FENNEL_LOCALS

FENNEL_LOCALS_CRATES_IO

FIDL_HIGHLIGHTS

FIDL_HIGHLIGHTS_CRATES_IO

FIDL_INJECTIONS

FIDL_INJECTIONS_CRATES_IO

FIDL_LOCALS

FIDL_LOCALS_CRATES_IO

FIRRTL_HIGHLIGHTS

FIRRTL_HIGHLIGHTS_CRATES_IO

FIRRTL_INJECTIONS

FIRRTL_INJECTIONS_CRATES_IO

FIRRTL_LOCALS

FIRRTL_LOCALS_CRATES_IO

FISH_HIGHLIGHTS

FISH_HIGHLIGHTS_CRATES_IO

FISH_INJECTIONS

FISH_INJECTIONS_CRATES_IO

FISH_LOCALS

FISH_LOCALS_CRATES_IO

FOAM_HIGHLIGHTS

FOAM_HIGHLIGHTS_CRATES_IO

FOAM_INJECTIONS

FOAM_INJECTIONS_CRATES_IO

FOAM_LOCALS

FOAM_LOCALS_CRATES_IO

FORTH_HIGHLIGHTS

FORTH_HIGHLIGHTS_CRATES_IO

FORTH_INJECTIONS

FORTH_INJECTIONS_CRATES_IO

FORTH_LOCALS

FORTH_LOCALS_CRATES_IO

FORTRAN_HIGHLIGHTS

FORTRAN_HIGHLIGHTS_CRATES_IO

FORTRAN_INJECTIONS

FORTRAN_INJECTIONS_CRATES_IO

FORTRAN_LOCALS

FORTRAN_LOCALS_CRATES_IO

FSD_HIGHLIGHTS

FSD_HIGHLIGHTS_CRATES_IO

FSD_INJECTIONS

FSD_INJECTIONS_CRATES_IO

FSD_LOCALS

FSD_LOCALS_CRATES_IO

FSHARP_HIGHLIGHTS

FSHARP_HIGHLIGHTS_CRATES_IO

FSHARP_INJECTIONS

FSHARP_INJECTIONS_CRATES_IO

FSHARP_LOCALS

FSHARP_LOCALS_CRATES_IO

FSH_HIGHLIGHTS

FSH_HIGHLIGHTS_CRATES_IO

FSH_INJECTIONS

FSH_INJECTIONS_CRATES_IO

FSH_LOCALS

FSH_LOCALS_CRATES_IO

FUNC_HIGHLIGHTS

FUNC_HIGHLIGHTS_CRATES_IO

FUNC_INJECTIONS

FUNC_INJECTIONS_CRATES_IO

FUNC_LOCALS

FUNC_LOCALS_CRATES_IO

FUSION_HIGHLIGHTS

FUSION_HIGHLIGHTS_CRATES_IO

FUSION_INJECTIONS

FUSION_INJECTIONS_CRATES_IO

FUSION_LOCALS

FUSION_LOCALS_CRATES_IO

GAPTST_HIGHLIGHTS

GAPTST_HIGHLIGHTS_CRATES_IO

GAPTST_INJECTIONS

GAPTST_INJECTIONS_CRATES_IO

GAPTST_LOCALS

GAPTST_LOCALS_CRATES_IO

GAP_HIGHLIGHTS

GAP_HIGHLIGHTS_CRATES_IO

GAP_INJECTIONS

GAP_INJECTIONS_CRATES_IO

GAP_LOCALS

GAP_LOCALS_CRATES_IO

GDRESOURCE_HIGHLIGHTS

GDRESOURCE_HIGHLIGHTS_CRATES_IO

GDRESOURCE_INJECTIONS

GDRESOURCE_INJECTIONS_CRATES_IO

GDRESOURCE_LOCALS

GDRESOURCE_LOCALS_CRATES_IO

GDSCRIPT_HIGHLIGHTS

GDSCRIPT_HIGHLIGHTS_CRATES_IO

GDSCRIPT_INJECTIONS

GDSCRIPT_INJECTIONS_CRATES_IO

GDSCRIPT_LOCALS

GDSCRIPT_LOCALS_CRATES_IO

GDSHADERINC_HIGHLIGHTS

GDSHADERINC_HIGHLIGHTS_CRATES_IO

GDSHADERINC_INJECTIONS

GDSHADERINC_INJECTIONS_CRATES_IO

GDSHADERINC_LOCALS

GDSHADERINC_LOCALS_CRATES_IO

GITATTRIBUTES_HIGHLIGHTS

GITATTRIBUTES_HIGHLIGHTS_CRATES_IO

GITATTRIBUTES_INJECTIONS

GITATTRIBUTES_INJECTIONS_CRATES_IO

GITATTRIBUTES_LOCALS

GITATTRIBUTES_LOCALS_CRATES_IO

GITCOMMIT_HIGHLIGHTS

GITCOMMIT_HIGHLIGHTS_CRATES_IO

GITCOMMIT_INJECTIONS

GITCOMMIT_INJECTIONS_CRATES_IO

GITCOMMIT_LOCALS

GITCOMMIT_LOCALS_CRATES_IO

GITCONFIG_HIGHLIGHTS

GITCONFIG_HIGHLIGHTS_CRATES_IO

GITCONFIG_INJECTIONS

GITCONFIG_INJECTIONS_CRATES_IO

GITCONFIG_LOCALS

GITCONFIG_LOCALS_CRATES_IO

GITDIFF_HIGHLIGHTS

GITDIFF_HIGHLIGHTS_CRATES_IO

GITDIFF_INJECTIONS

GITDIFF_INJECTIONS_CRATES_IO

GITDIFF_LOCALS

GITDIFF_LOCALS_CRATES_IO

GITIGNORE_HIGHLIGHTS

GITIGNORE_HIGHLIGHTS_CRATES_IO

GITIGNORE_INJECTIONS

GITIGNORE_INJECTIONS_CRATES_IO

GITIGNORE_LOCALS

GITIGNORE_LOCALS_CRATES_IO

GITREBASE_HIGHLIGHTS

GITREBASE_HIGHLIGHTS_CRATES_IO

GITREBASE_INJECTIONS

GITREBASE_INJECTIONS_CRATES_IO

GITREBASE_LOCALS

GITREBASE_LOCALS_CRATES_IO

GLEAM_HIGHLIGHTS

GLEAM_HIGHLIGHTS_CRATES_IO

GLEAM_INJECTIONS

GLEAM_INJECTIONS_CRATES_IO

GLEAM_LOCALS

GLEAM_LOCALS_CRATES_IO

GLIMMER_JAVASCRIPT_HIGHLIGHTS

GLIMMER_JAVASCRIPT_HIGHLIGHTS_CRATES_IO

GLIMMER_JAVASCRIPT_INJECTIONS

GLIMMER_JAVASCRIPT_INJECTIONS_CRATES_IO

GLIMMER_JAVASCRIPT_LOCALS

GLIMMER_JAVASCRIPT_LOCALS_CRATES_IO

GLIMMER_TYPESCRIPT_HIGHLIGHTS

GLIMMER_TYPESCRIPT_HIGHLIGHTS_CRATES_IO

GLIMMER_TYPESCRIPT_INJECTIONS

GLIMMER_TYPESCRIPT_INJECTIONS_CRATES_IO

GLIMMER_TYPESCRIPT_LOCALS

GLIMMER_TYPESCRIPT_LOCALS_CRATES_IO

GLSL_HIGHLIGHTS

GLSL_HIGHLIGHTS_CRATES_IO

GLSL_INJECTIONS

GLSL_INJECTIONS_CRATES_IO

GLSL_LOCALS

GLSL_LOCALS_CRATES_IO

GNUPLOT_HIGHLIGHTS

GNUPLOT_HIGHLIGHTS_CRATES_IO

GNUPLOT_INJECTIONS

GNUPLOT_INJECTIONS_CRATES_IO

GNUPLOT_LOCALS

GNUPLOT_LOCALS_CRATES_IO

GN_HIGHLIGHTS

GN_HIGHLIGHTS_CRATES_IO

GN_INJECTIONS

GN_INJECTIONS_CRATES_IO

GN_LOCALS

GN_LOCALS_CRATES_IO

GOCTL_HIGHLIGHTS

GOCTL_HIGHLIGHTS_CRATES_IO

GOCTL_INJECTIONS

GOCTL_INJECTIONS_CRATES_IO

GOCTL_LOCALS

GOCTL_LOCALS_CRATES_IO

GOMOD_HIGHLIGHTS

GOMOD_HIGHLIGHTS_CRATES_IO

GOMOD_INJECTIONS

GOMOD_INJECTIONS_CRATES_IO

GOMOD_LOCALS

GOMOD_LOCALS_CRATES_IO

GOSUM_HIGHLIGHTS

GOSUM_HIGHLIGHTS_CRATES_IO

GOSUM_INJECTIONS

GOSUM_INJECTIONS_CRATES_IO

GOSUM_LOCALS

GOSUM_LOCALS_CRATES_IO

GOTMPL_HIGHLIGHTS

GOTMPL_HIGHLIGHTS_CRATES_IO

GOTMPL_INJECTIONS

GOTMPL_INJECTIONS_CRATES_IO

GOTMPL_LOCALS

GOTMPL_LOCALS_CRATES_IO

GOWORK_HIGHLIGHTS

GOWORK_HIGHLIGHTS_CRATES_IO

GOWORK_INJECTIONS

GOWORK_INJECTIONS_CRATES_IO

GOWORK_LOCALS

GOWORK_LOCALS_CRATES_IO

GO_HIGHLIGHTS

GO_HIGHLIGHTS_CRATES_IO

GO_INJECTIONS

GO_INJECTIONS_CRATES_IO

GO_LOCALS

GO_LOCALS_CRATES_IO

GPG_HIGHLIGHTS

GPG_HIGHLIGHTS_CRATES_IO

GPG_INJECTIONS

GPG_INJECTIONS_CRATES_IO

GPG_LOCALS

GPG_LOCALS_CRATES_IO

GRAPHQL_HIGHLIGHTS

GRAPHQL_HIGHLIGHTS_CRATES_IO

GRAPHQL_INJECTIONS

GRAPHQL_INJECTIONS_CRATES_IO

GRAPHQL_LOCALS

GRAPHQL_LOCALS_CRATES_IO

GREN_HIGHLIGHTS

GREN_HIGHLIGHTS_CRATES_IO

GREN_INJECTIONS

GREN_INJECTIONS_CRATES_IO

GREN_LOCALS

GREN_LOCALS_CRATES_IO

GROOVY_HIGHLIGHTS

GROOVY_HIGHLIGHTS_CRATES_IO

GROOVY_INJECTIONS

GROOVY_INJECTIONS_CRATES_IO

GROOVY_LOCALS

GROOVY_LOCALS_CRATES_IO

GROQ_HIGHLIGHTS

GROQ_HIGHLIGHTS_CRATES_IO

GROQ_INJECTIONS

GROQ_INJECTIONS_CRATES_IO

GROQ_LOCALS

GROQ_LOCALS_CRATES_IO

GSTLAUNCH_HIGHLIGHTS

GSTLAUNCH_HIGHLIGHTS_CRATES_IO

GSTLAUNCH_INJECTIONS

GSTLAUNCH_INJECTIONS_CRATES_IO

GSTLAUNCH_LOCALS

GSTLAUNCH_LOCALS_CRATES_IO

HACK_HIGHLIGHTS

HACK_HIGHLIGHTS_CRATES_IO

HACK_INJECTIONS

HACK_INJECTIONS_CRATES_IO

HACK_LOCALS

HACK_LOCALS_CRATES_IO

HANDLEBARS_HIGHLIGHTS

HANDLEBARS_HIGHLIGHTS_CRATES_IO

HANDLEBARS_INJECTIONS

HANDLEBARS_INJECTIONS_CRATES_IO

HANDLEBARS_LOCALS

HANDLEBARS_LOCALS_CRATES_IO

HARE_HIGHLIGHTS

HARE_HIGHLIGHTS_CRATES_IO

HARE_INJECTIONS

HARE_INJECTIONS_CRATES_IO

HARE_LOCALS

HARE_LOCALS_CRATES_IO

HASKELLPERSISTENT_HIGHLIGHTS

HASKELLPERSISTENT_HIGHLIGHTS_CRATES_IO

HASKELLPERSISTENT_INJECTIONS

HASKELLPERSISTENT_INJECTIONS_CRATES_IO

HASKELLPERSISTENT_LOCALS

HASKELLPERSISTENT_LOCALS_CRATES_IO

HASKELL_HIGHLIGHTS

HASKELL_HIGHLIGHTS_CRATES_IO

HASKELL_INJECTIONS

HASKELL_INJECTIONS_CRATES_IO

HASKELL_LOCALS

HASKELL_LOCALS_CRATES_IO

HAXE_HIGHLIGHTS

HAXE_HIGHLIGHTS_CRATES_IO

HAXE_INJECTIONS

HAXE_INJECTIONS_CRATES_IO

HAXE_LOCALS

HAXE_LOCALS_CRATES_IO

HCL_HIGHLIGHTS

HCL_HIGHLIGHTS_CRATES_IO

HCL_INJECTIONS

HCL_INJECTIONS_CRATES_IO

HCL_LOCALS

HCL_LOCALS_CRATES_IO

HEEX_HIGHLIGHTS

HEEX_HIGHLIGHTS_CRATES_IO

HEEX_INJECTIONS

HEEX_INJECTIONS_CRATES_IO

HEEX_LOCALS

HEEX_LOCALS_CRATES_IO

HELM_HIGHLIGHTS

HELM_HIGHLIGHTS_CRATES_IO

HELM_INJECTIONS

HELM_INJECTIONS_CRATES_IO

HELM_LOCALS

HELM_LOCALS_CRATES_IO

HELP_HIGHLIGHTS

HELP_HIGHLIGHTS_CRATES_IO

HELP_INJECTIONS

HELP_INJECTIONS_CRATES_IO

HELP_LOCALS

HELP_LOCALS_CRATES_IO

HEXDUMP_HIGHLIGHTS

HEXDUMP_HIGHLIGHTS_CRATES_IO

HEXDUMP_INJECTIONS

HEXDUMP_INJECTIONS_CRATES_IO

HEXDUMP_LOCALS

HEXDUMP_LOCALS_CRATES_IO

HJSON_HIGHLIGHTS

HJSON_HIGHLIGHTS_CRATES_IO

HJSON_INJECTIONS

HJSON_INJECTIONS_CRATES_IO

HJSON_LOCALS

HJSON_LOCALS_CRATES_IO

HLSL_HIGHLIGHTS

HLSL_HIGHLIGHTS_CRATES_IO

HLSL_INJECTIONS

HLSL_INJECTIONS_CRATES_IO

HLSL_LOCALS

HLSL_LOCALS_CRATES_IO

HLSPLAYLIST_HIGHLIGHTS

HLSPLAYLIST_HIGHLIGHTS_CRATES_IO

HLSPLAYLIST_INJECTIONS

HLSPLAYLIST_INJECTIONS_CRATES_IO

HLSPLAYLIST_LOCALS

HLSPLAYLIST_LOCALS_CRATES_IO

HOCON_HIGHLIGHTS

HOCON_HIGHLIGHTS_CRATES_IO

HOCON_INJECTIONS

HOCON_INJECTIONS_CRATES_IO

HOCON_LOCALS

HOCON_LOCALS_CRATES_IO

HOON_HIGHLIGHTS

HOON_HIGHLIGHTS_CRATES_IO

HOON_INJECTIONS

HOON_INJECTIONS_CRATES_IO

HOON_LOCALS

HOON_LOCALS_CRATES_IO

HTMLANGULAR_HIGHLIGHTS

HTMLANGULAR_HIGHLIGHTS_CRATES_IO

HTMLANGULAR_INJECTIONS

HTMLANGULAR_INJECTIONS_CRATES_IO

HTMLANGULAR_LOCALS

HTMLANGULAR_LOCALS_CRATES_IO

HTMLDJANGO_HIGHLIGHTS

HTMLDJANGO_HIGHLIGHTS_CRATES_IO

HTMLDJANGO_INJECTIONS

HTMLDJANGO_INJECTIONS_CRATES_IO

HTMLDJANGO_LOCALS

HTMLDJANGO_LOCALS_CRATES_IO

HTML_HIGHLIGHTS

HTML_HIGHLIGHTS_CRATES_IO

HTML_INJECTIONS

HTML_INJECTIONS_CRATES_IO

HTML_LOCALS

HTML_LOCALS_CRATES_IO

HTTP_HIGHLIGHTS

HTTP_HIGHLIGHTS_CRATES_IO

HTTP_INJECTIONS

HTTP_INJECTIONS_CRATES_IO

HTTP_LOCALS

HTTP_LOCALS_CRATES_IO

HURL_HIGHLIGHTS

HURL_HIGHLIGHTS_CRATES_IO

HURL_INJECTIONS

HURL_INJECTIONS_CRATES_IO

HURL_LOCALS

HURL_LOCALS_CRATES_IO

HYPRLANG_HIGHLIGHTS

HYPRLANG_HIGHLIGHTS_CRATES_IO

HYPRLANG_INJECTIONS

HYPRLANG_INJECTIONS_CRATES_IO

HYPRLANG_LOCALS

HYPRLANG_LOCALS_CRATES_IO

IDL_HIGHLIGHTS

IDL_HIGHLIGHTS_CRATES_IO

IDL_INJECTIONS

IDL_INJECTIONS_CRATES_IO

IDL_LOCALS

IDL_LOCALS_CRATES_IO

IDRIS2_HIGHLIGHTS

IDRIS2_HIGHLIGHTS_CRATES_IO

IDRIS2_INJECTIONS

IDRIS2_INJECTIONS_CRATES_IO

IDRIS2_LOCALS

IDRIS2_LOCALS_CRATES_IO

INI_HIGHLIGHTS

INI_HIGHLIGHTS_CRATES_IO

INI_INJECTIONS

INI_INJECTIONS_CRATES_IO

INI_LOCALS

INI_LOCALS_CRATES_IO

INKO_HIGHLIGHTS

INKO_HIGHLIGHTS_CRATES_IO

INKO_INJECTIONS

INKO_INJECTIONS_CRATES_IO

INKO_LOCALS

INKO_LOCALS_CRATES_IO

ISPC_HIGHLIGHTS

ISPC_HIGHLIGHTS_CRATES_IO

ISPC_INJECTIONS

ISPC_INJECTIONS_CRATES_IO

ISPC_LOCALS

ISPC_LOCALS_CRATES_IO

JANET_HIGHLIGHTS

JANET_HIGHLIGHTS_CRATES_IO

JANET_INJECTIONS

JANET_INJECTIONS_CRATES_IO

JANET_LOCALS

JANET_LOCALS_CRATES_IO

JAVADOC_HIGHLIGHTS

JAVADOC_HIGHLIGHTS_CRATES_IO

JAVADOC_INJECTIONS

JAVADOC_INJECTIONS_CRATES_IO

JAVADOC_LOCALS

JAVADOC_LOCALS_CRATES_IO

JAVASCRIPT_HIGHLIGHTS

JAVASCRIPT_HIGHLIGHTS_CRATES_IO

JAVASCRIPT_INJECTIONS

JAVASCRIPT_INJECTIONS_CRATES_IO

JAVASCRIPT_LOCALS

JAVASCRIPT_LOCALS_CRATES_IO

JAVA_HIGHLIGHTS

JAVA_HIGHLIGHTS_CRATES_IO

JAVA_INJECTIONS

JAVA_INJECTIONS_CRATES_IO

JAVA_LOCALS

JAVA_LOCALS_CRATES_IO

JINJA_HIGHLIGHTS

JINJA_HIGHLIGHTS_CRATES_IO

JINJA_INJECTIONS

JINJA_INJECTIONS_CRATES_IO

JINJA_INLINE_HIGHLIGHTS

JINJA_INLINE_HIGHLIGHTS_CRATES_IO

JINJA_INLINE_INJECTIONS

JINJA_INLINE_INJECTIONS_CRATES_IO

JINJA_INLINE_LOCALS

JINJA_INLINE_LOCALS_CRATES_IO

JINJA_LOCALS

JINJA_LOCALS_CRATES_IO

JPROPERTIES_HIGHLIGHTS

JPROPERTIES_HIGHLIGHTS_CRATES_IO

JPROPERTIES_INJECTIONS

JPROPERTIES_INJECTIONS_CRATES_IO

JPROPERTIES_LOCALS

JPROPERTIES_LOCALS_CRATES_IO

JQ_HIGHLIGHTS

JQ_HIGHLIGHTS_CRATES_IO

JQ_INJECTIONS

JQ_INJECTIONS_CRATES_IO

JQ_LOCALS

JQ_LOCALS_CRATES_IO

JSDOC_HIGHLIGHTS

JSDOC_HIGHLIGHTS_CRATES_IO

JSDOC_INJECTIONS

JSDOC_INJECTIONS_CRATES_IO

JSDOC_LOCALS

JSDOC_LOCALS_CRATES_IO

JSON5_HIGHLIGHTS

JSON5_HIGHLIGHTS_CRATES_IO

JSON5_INJECTIONS

JSON5_INJECTIONS_CRATES_IO

JSON5_LOCALS

JSON5_LOCALS_CRATES_IO

JSONC_HIGHLIGHTS

JSONC_HIGHLIGHTS_CRATES_IO

JSONC_INJECTIONS

JSONC_INJECTIONS_CRATES_IO

JSONC_LOCALS

JSONC_LOCALS_CRATES_IO

JSONNET_HIGHLIGHTS

JSONNET_HIGHLIGHTS_CRATES_IO

JSONNET_INJECTIONS

JSONNET_INJECTIONS_CRATES_IO

JSONNET_LOCALS

JSONNET_LOCALS_CRATES_IO

JSON_HIGHLIGHTS

JSON_HIGHLIGHTS_CRATES_IO

JSON_INJECTIONS

JSON_INJECTIONS_CRATES_IO

JSON_LOCALS

JSON_LOCALS_CRATES_IO

JULIA_HIGHLIGHTS

JULIA_HIGHLIGHTS_CRATES_IO

JULIA_INJECTIONS

JULIA_INJECTIONS_CRATES_IO

JULIA_LOCALS

JULIA_LOCALS_CRATES_IO

JUST_HIGHLIGHTS

JUST_HIGHLIGHTS_CRATES_IO

JUST_INJECTIONS

JUST_INJECTIONS_CRATES_IO

JUST_LOCALS

JUST_LOCALS_CRATES_IO

KCL_HIGHLIGHTS

KCL_HIGHLIGHTS_CRATES_IO

KCL_INJECTIONS

KCL_INJECTIONS_CRATES_IO

KCL_LOCALS

KCL_LOCALS_CRATES_IO

KCONFIG_HIGHLIGHTS

KCONFIG_HIGHLIGHTS_CRATES_IO

KCONFIG_INJECTIONS

KCONFIG_INJECTIONS_CRATES_IO

KCONFIG_LOCALS

KCONFIG_LOCALS_CRATES_IO

KDL_HIGHLIGHTS

KDL_HIGHLIGHTS_CRATES_IO

KDL_INJECTIONS

KDL_INJECTIONS_CRATES_IO

KDL_LOCALS

KDL_LOCALS_CRATES_IO

KITTY_HIGHLIGHTS

KITTY_HIGHLIGHTS_CRATES_IO

KITTY_INJECTIONS

KITTY_INJECTIONS_CRATES_IO

KITTY_LOCALS

KITTY_LOCALS_CRATES_IO

KOS_HIGHLIGHTS

KOS_HIGHLIGHTS_CRATES_IO

KOS_INJECTIONS

KOS_INJECTIONS_CRATES_IO

KOS_LOCALS

KOS_LOCALS_CRATES_IO

KOTLIN_HIGHLIGHTS

KOTLIN_HIGHLIGHTS_CRATES_IO

KOTLIN_INJECTIONS

KOTLIN_INJECTIONS_CRATES_IO

KOTLIN_LOCALS

KOTLIN_LOCALS_CRATES_IO

KOTO_HIGHLIGHTS

KOTO_HIGHLIGHTS_CRATES_IO

KOTO_INJECTIONS

KOTO_INJECTIONS_CRATES_IO

KOTO_LOCALS

KOTO_LOCALS_CRATES_IO

KUSTO_HIGHLIGHTS

KUSTO_HIGHLIGHTS_CRATES_IO

KUSTO_INJECTIONS

KUSTO_INJECTIONS_CRATES_IO

KUSTO_LOCALS

KUSTO_LOCALS_CRATES_IO

LALRPOP_HIGHLIGHTS

LALRPOP_HIGHLIGHTS_CRATES_IO

LALRPOP_INJECTIONS

LALRPOP_INJECTIONS_CRATES_IO

LALRPOP_LOCALS

LALRPOP_LOCALS_CRATES_IO

LATEX_HIGHLIGHTS

LATEX_HIGHLIGHTS_CRATES_IO

LATEX_INJECTIONS

LATEX_INJECTIONS_CRATES_IO

LATEX_LOCALS

LATEX_LOCALS_CRATES_IO

LD_HIGHLIGHTS

LD_HIGHLIGHTS_CRATES_IO

LD_INJECTIONS

LD_INJECTIONS_CRATES_IO

LD_LOCALS

LD_LOCALS_CRATES_IO

LEDGER_HIGHLIGHTS

LEDGER_HIGHLIGHTS_CRATES_IO

LEDGER_INJECTIONS

LEDGER_INJECTIONS_CRATES_IO

LEDGER_LOCALS

LEDGER_LOCALS_CRATES_IO

LEO_HIGHLIGHTS

LEO_HIGHLIGHTS_CRATES_IO

LEO_INJECTIONS

LEO_INJECTIONS_CRATES_IO

LEO_LOCALS

LEO_LOCALS_CRATES_IO

LIQUIDSOAP_HIGHLIGHTS

LIQUIDSOAP_HIGHLIGHTS_CRATES_IO

LIQUIDSOAP_INJECTIONS

LIQUIDSOAP_INJECTIONS_CRATES_IO

LIQUIDSOAP_LOCALS

LIQUIDSOAP_LOCALS_CRATES_IO

LIQUID_HIGHLIGHTS

LIQUID_HIGHLIGHTS_CRATES_IO

LIQUID_INJECTIONS

LIQUID_INJECTIONS_CRATES_IO

LIQUID_LOCALS

LIQUID_LOCALS_CRATES_IO

LISP_HIGHLIGHTS

LISP_HIGHLIGHTS_CRATES_IO

LISP_INJECTIONS

LISP_INJECTIONS_CRATES_IO

LISP_LOCALS

LISP_LOCALS_CRATES_IO

LLVM_HIGHLIGHTS

LLVM_HIGHLIGHTS_CRATES_IO

LLVM_INJECTIONS

LLVM_INJECTIONS_CRATES_IO

LLVM_LOCALS

LLVM_LOCALS_CRATES_IO

LUADOC_HIGHLIGHTS

LUADOC_HIGHLIGHTS_CRATES_IO

LUADOC_INJECTIONS

LUADOC_INJECTIONS_CRATES_IO

LUADOC_LOCALS

LUADOC_LOCALS_CRATES_IO

LUAP_HIGHLIGHTS

LUAP_HIGHLIGHTS_CRATES_IO

LUAP_INJECTIONS

LUAP_INJECTIONS_CRATES_IO

LUAP_LOCALS

LUAP_LOCALS_CRATES_IO

LUAU_HIGHLIGHTS

LUAU_HIGHLIGHTS_CRATES_IO

LUAU_INJECTIONS

LUAU_INJECTIONS_CRATES_IO

LUAU_LOCALS

LUAU_LOCALS_CRATES_IO

LUA_HIGHLIGHTS

LUA_HIGHLIGHTS_CRATES_IO

LUA_INJECTIONS

LUA_INJECTIONS_CRATES_IO

LUA_LOCALS

LUA_LOCALS_CRATES_IO

MAGIK_HIGHLIGHTS

MAGIK_HIGHLIGHTS_CRATES_IO

MAGIK_INJECTIONS

MAGIK_INJECTIONS_CRATES_IO

MAGIK_LOCALS

MAGIK_LOCALS_CRATES_IO

MAKE_HIGHLIGHTS

MAKE_HIGHLIGHTS_CRATES_IO

MAKE_INJECTIONS

MAKE_INJECTIONS_CRATES_IO

MAKE_LOCALS

MAKE_LOCALS_CRATES_IO

MARKDOWN_HIGHLIGHTS

MARKDOWN_HIGHLIGHTS_CRATES_IO

MARKDOWN_INJECTIONS

MARKDOWN_INJECTIONS_CRATES_IO

MARKDOWN_INLINE_HIGHLIGHTS

MARKDOWN_INLINE_HIGHLIGHTS_CRATES_IO

MARKDOWN_INLINE_INJECTIONS

MARKDOWN_INLINE_INJECTIONS_CRATES_IO

MARKDOWN_INLINE_LOCALS

MARKDOWN_INLINE_LOCALS_CRATES_IO

MARKDOWN_LOCALS

MARKDOWN_LOCALS_CRATES_IO

MATLAB_HIGHLIGHTS

MATLAB_HIGHLIGHTS_CRATES_IO

MATLAB_INJECTIONS

MATLAB_INJECTIONS_CRATES_IO

MATLAB_LOCALS

MATLAB_LOCALS_CRATES_IO

MENHIR_HIGHLIGHTS

MENHIR_HIGHLIGHTS_CRATES_IO

MENHIR_INJECTIONS

MENHIR_INJECTIONS_CRATES_IO

MENHIR_LOCALS

MENHIR_LOCALS_CRATES_IO

MERMAID_HIGHLIGHTS

MERMAID_HIGHLIGHTS_CRATES_IO

MERMAID_INJECTIONS

MERMAID_INJECTIONS_CRATES_IO

MERMAID_LOCALS

MERMAID_LOCALS_CRATES_IO

MESON_HIGHLIGHTS

MESON_HIGHLIGHTS_CRATES_IO

MESON_INJECTIONS

MESON_INJECTIONS_CRATES_IO

MESON_LOCALS

MESON_LOCALS_CRATES_IO

MLIR_HIGHLIGHTS

MLIR_HIGHLIGHTS_CRATES_IO

MLIR_INJECTIONS

MLIR_INJECTIONS_CRATES_IO

MLIR_LOCALS

MLIR_LOCALS_CRATES_IO

MUTTRC_HIGHLIGHTS

MUTTRC_HIGHLIGHTS_CRATES_IO

MUTTRC_INJECTIONS

MUTTRC_INJECTIONS_CRATES_IO

MUTTRC_LOCALS

MUTTRC_LOCALS_CRATES_IO

NASM_HIGHLIGHTS

NASM_HIGHLIGHTS_CRATES_IO

NASM_INJECTIONS

NASM_INJECTIONS_CRATES_IO

NASM_LOCALS

NASM_LOCALS_CRATES_IO

NGINX_HIGHLIGHTS

NGINX_HIGHLIGHTS_CRATES_IO

NGINX_INJECTIONS

NGINX_INJECTIONS_CRATES_IO

NGINX_LOCALS

NGINX_LOCALS_CRATES_IO

NICKEL_HIGHLIGHTS

NICKEL_HIGHLIGHTS_CRATES_IO

NICKEL_INJECTIONS

NICKEL_INJECTIONS_CRATES_IO

NICKEL_LOCALS

NICKEL_LOCALS_CRATES_IO

NIM_FORMAT_STRING_HIGHLIGHTS

NIM_FORMAT_STRING_HIGHLIGHTS_CRATES_IO

NIM_FORMAT_STRING_INJECTIONS

NIM_FORMAT_STRING_INJECTIONS_CRATES_IO

NIM_FORMAT_STRING_LOCALS

NIM_FORMAT_STRING_LOCALS_CRATES_IO

NIM_HIGHLIGHTS

NIM_HIGHLIGHTS_CRATES_IO

NIM_INJECTIONS

NIM_INJECTIONS_CRATES_IO

NIM_LOCALS

NIM_LOCALS_CRATES_IO

NINJA_HIGHLIGHTS

NINJA_HIGHLIGHTS_CRATES_IO

NINJA_INJECTIONS

NINJA_INJECTIONS_CRATES_IO

NINJA_LOCALS

NINJA_LOCALS_CRATES_IO

NIX_HIGHLIGHTS

NIX_HIGHLIGHTS_CRATES_IO

NIX_INJECTIONS

NIX_INJECTIONS_CRATES_IO

NIX_LOCALS

NIX_LOCALS_CRATES_IO

NQC_HIGHLIGHTS

NQC_HIGHLIGHTS_CRATES_IO

NQC_INJECTIONS

NQC_INJECTIONS_CRATES_IO

NQC_LOCALS

NQC_LOCALS_CRATES_IO

NU_HIGHLIGHTS

NU_HIGHLIGHTS_CRATES_IO

NU_INJECTIONS

NU_INJECTIONS_CRATES_IO

NU_LOCALS

NU_LOCALS_CRATES_IO

OBJC_HIGHLIGHTS

OBJC_HIGHLIGHTS_CRATES_IO

OBJC_INJECTIONS

OBJC_INJECTIONS_CRATES_IO

OBJC_LOCALS

OBJC_LOCALS_CRATES_IO

OBJDUMP_HIGHLIGHTS

OBJDUMP_HIGHLIGHTS_CRATES_IO

OBJDUMP_INJECTIONS

OBJDUMP_INJECTIONS_CRATES_IO

OBJDUMP_LOCALS

OBJDUMP_LOCALS_CRATES_IO

OCAMLINTERFACE_HIGHLIGHTS

OCAMLINTERFACE_HIGHLIGHTS_CRATES_IO

OCAMLINTERFACE_INJECTIONS

OCAMLINTERFACE_INJECTIONS_CRATES_IO

OCAMLINTERFACE_LOCALS

OCAMLINTERFACE_LOCALS_CRATES_IO

OCAMLLEX_HIGHLIGHTS

OCAMLLEX_HIGHLIGHTS_CRATES_IO

OCAMLLEX_INJECTIONS

OCAMLLEX_INJECTIONS_CRATES_IO

OCAMLLEX_LOCALS

OCAMLLEX_LOCALS_CRATES_IO

OCAML_HIGHLIGHTS

OCAML_HIGHLIGHTS_CRATES_IO

OCAML_INJECTIONS

OCAML_INJECTIONS_CRATES_IO

OCAML_INTERFACE_HIGHLIGHTS

OCAML_INTERFACE_HIGHLIGHTS_CRATES_IO

OCAML_INTERFACE_INJECTIONS

OCAML_INTERFACE_INJECTIONS_CRATES_IO

OCAML_INTERFACE_LOCALS

OCAML_INTERFACE_LOCALS_CRATES_IO

OCAML_LOCALS

OCAML_LOCALS_CRATES_IO

ODIN_HIGHLIGHTS

ODIN_HIGHLIGHTS_CRATES_IO

ODIN_INJECTIONS

ODIN_INJECTIONS_CRATES_IO

ODIN_LOCALS

ODIN_LOCALS_CRATES_IO

PASCAL_HIGHLIGHTS

PASCAL_HIGHLIGHTS_CRATES_IO

PASCAL_INJECTIONS

PASCAL_INJECTIONS_CRATES_IO

PASCAL_LOCALS

PASCAL_LOCALS_CRATES_IO

PASSWD_HIGHLIGHTS

PASSWD_HIGHLIGHTS_CRATES_IO

PASSWD_INJECTIONS

PASSWD_INJECTIONS_CRATES_IO

PASSWD_LOCALS

PASSWD_LOCALS_CRATES_IO

PBTXT_HIGHLIGHTS

PBTXT_HIGHLIGHTS_CRATES_IO

PBTXT_INJECTIONS

PBTXT_INJECTIONS_CRATES_IO

PBTXT_LOCALS

PBTXT_LOCALS_CRATES_IO

PEM_HIGHLIGHTS

PEM_HIGHLIGHTS_CRATES_IO

PEM_INJECTIONS

PEM_INJECTIONS_CRATES_IO

PEM_LOCALS

PEM_LOCALS_CRATES_IO

PERL_HIGHLIGHTS

PERL_HIGHLIGHTS_CRATES_IO

PERL_INJECTIONS

PERL_INJECTIONS_CRATES_IO

PERL_LOCALS

PERL_LOCALS_CRATES_IO

PGN_HIGHLIGHTS

PGN_HIGHLIGHTS_CRATES_IO

PGN_INJECTIONS

PGN_INJECTIONS_CRATES_IO

PGN_LOCALS

PGN_LOCALS_CRATES_IO

PHPDOC_HIGHLIGHTS

PHPDOC_HIGHLIGHTS_CRATES_IO

PHPDOC_INJECTIONS

PHPDOC_INJECTIONS_CRATES_IO

PHPDOC_LOCALS

PHPDOC_LOCALS_CRATES_IO

PHP_HIGHLIGHTS

PHP_HIGHLIGHTS_CRATES_IO

PHP_INJECTIONS

PHP_INJECTIONS_CRATES_IO

PHP_LOCALS

PHP_LOCALS_CRATES_IO

PHP_ONLY_HIGHLIGHTS

PHP_ONLY_HIGHLIGHTS_CRATES_IO

PHP_ONLY_INJECTIONS

PHP_ONLY_INJECTIONS_CRATES_IO

PHP_ONLY_LOCALS

PHP_ONLY_LOCALS_CRATES_IO

PIOASM_HIGHLIGHTS

PIOASM_HIGHLIGHTS_CRATES_IO

PIOASM_INJECTIONS

PIOASM_INJECTIONS_CRATES_IO

PIOASM_LOCALS

PIOASM_LOCALS_CRATES_IO

PKL_HIGHLIGHTS

PKL_HIGHLIGHTS_CRATES_IO

PKL_INJECTIONS

PKL_INJECTIONS_CRATES_IO

PKL_LOCALS

PKL_LOCALS_CRATES_IO

POD_HIGHLIGHTS

POD_HIGHLIGHTS_CRATES_IO

POD_INJECTIONS

POD_INJECTIONS_CRATES_IO

POD_LOCALS

POD_LOCALS_CRATES_IO

POEFILTER_HIGHLIGHTS

POEFILTER_HIGHLIGHTS_CRATES_IO

POEFILTER_INJECTIONS

POEFILTER_INJECTIONS_CRATES_IO

POEFILTER_LOCALS

POEFILTER_LOCALS_CRATES_IO

PONY_HIGHLIGHTS

PONY_HIGHLIGHTS_CRATES_IO

PONY_INJECTIONS

PONY_INJECTIONS_CRATES_IO

PONY_LOCALS

PONY_LOCALS_CRATES_IO

PO_HIGHLIGHTS

PO_HIGHLIGHTS_CRATES_IO

PO_INJECTIONS

PO_INJECTIONS_CRATES_IO

PO_LOCALS

PO_LOCALS_CRATES_IO

PRINTF_HIGHLIGHTS

PRINTF_HIGHLIGHTS_CRATES_IO

PRINTF_INJECTIONS

PRINTF_INJECTIONS_CRATES_IO

PRINTF_LOCALS

PRINTF_LOCALS_CRATES_IO

PRISMA_HIGHLIGHTS

PRISMA_HIGHLIGHTS_CRATES_IO

PRISMA_INJECTIONS

PRISMA_INJECTIONS_CRATES_IO

PRISMA_LOCALS

PRISMA_LOCALS_CRATES_IO

PROBLOG_HIGHLIGHTS

PROBLOG_HIGHLIGHTS_CRATES_IO

PROBLOG_INJECTIONS

PROBLOG_INJECTIONS_CRATES_IO

PROBLOG_LOCALS

PROBLOG_LOCALS_CRATES_IO

PROLOG_HIGHLIGHTS

PROLOG_HIGHLIGHTS_CRATES_IO

PROLOG_INJECTIONS

PROLOG_INJECTIONS_CRATES_IO

PROLOG_LOCALS

PROLOG_LOCALS_CRATES_IO

PROMQL_HIGHLIGHTS

PROMQL_HIGHLIGHTS_CRATES_IO

PROMQL_INJECTIONS

PROMQL_INJECTIONS_CRATES_IO

PROMQL_LOCALS

PROMQL_LOCALS_CRATES_IO

PROTO_HIGHLIGHTS

PROTO_HIGHLIGHTS_CRATES_IO

PROTO_INJECTIONS

PROTO_INJECTIONS_CRATES_IO

PROTO_LOCALS

PROTO_LOCALS_CRATES_IO

PRQL_HIGHLIGHTS

PRQL_HIGHLIGHTS_CRATES_IO

PRQL_INJECTIONS

PRQL_INJECTIONS_CRATES_IO

PRQL_LOCALS

PRQL_LOCALS_CRATES_IO

PS1_HIGHLIGHTS

PS1_HIGHLIGHTS_CRATES_IO

PS1_INJECTIONS

PS1_INJECTIONS_CRATES_IO

PS1_LOCALS

PS1_LOCALS_CRATES_IO

PSV_HIGHLIGHTS

PSV_HIGHLIGHTS_CRATES_IO

PSV_INJECTIONS

PSV_INJECTIONS_CRATES_IO

PSV_LOCALS

PSV_LOCALS_CRATES_IO

PUG_HIGHLIGHTS

PUG_HIGHLIGHTS_CRATES_IO

PUG_INJECTIONS

PUG_INJECTIONS_CRATES_IO

PUG_LOCALS

PUG_LOCALS_CRATES_IO

PUPPET_HIGHLIGHTS

PUPPET_HIGHLIGHTS_CRATES_IO

PUPPET_INJECTIONS

PUPPET_INJECTIONS_CRATES_IO

PUPPET_LOCALS

PUPPET_LOCALS_CRATES_IO

PURESCRIPT_HIGHLIGHTS

PURESCRIPT_HIGHLIGHTS_CRATES_IO

PURESCRIPT_INJECTIONS

PURESCRIPT_INJECTIONS_CRATES_IO

PURESCRIPT_LOCALS

PURESCRIPT_LOCALS_CRATES_IO

PYMANIFEST_HIGHLIGHTS

PYMANIFEST_HIGHLIGHTS_CRATES_IO

PYMANIFEST_INJECTIONS

PYMANIFEST_INJECTIONS_CRATES_IO

PYMANIFEST_LOCALS

PYMANIFEST_LOCALS_CRATES_IO

PYTHON_HIGHLIGHTS

PYTHON_HIGHLIGHTS_CRATES_IO

PYTHON_INJECTIONS

PYTHON_INJECTIONS_CRATES_IO

PYTHON_LOCALS

PYTHON_LOCALS_CRATES_IO

QL_HIGHLIGHTS

QL_HIGHLIGHTS_CRATES_IO

QL_INJECTIONS

QL_INJECTIONS_CRATES_IO

QL_LOCALS

QL_LOCALS_CRATES_IO

QMLDIR_HIGHLIGHTS

QMLDIR_HIGHLIGHTS_CRATES_IO

QMLDIR_INJECTIONS

QMLDIR_INJECTIONS_CRATES_IO

QMLDIR_LOCALS

QMLDIR_LOCALS_CRATES_IO

QML_HIGHLIGHTS

QML_HIGHLIGHTS_CRATES_IO

QML_INJECTIONS

QML_INJECTIONS_CRATES_IO

QML_LOCALS

QML_LOCALS_CRATES_IO

QUERY_HIGHLIGHTS

QUERY_HIGHLIGHTS_CRATES_IO

QUERY_INJECTIONS

QUERY_INJECTIONS_CRATES_IO

QUERY_LOCALS

QUERY_LOCALS_CRATES_IO

RACKET_HIGHLIGHTS

RACKET_HIGHLIGHTS_CRATES_IO

RACKET_INJECTIONS

RACKET_INJECTIONS_CRATES_IO

RACKET_LOCALS

RACKET_LOCALS_CRATES_IO

RALPH_HIGHLIGHTS

RALPH_HIGHLIGHTS_CRATES_IO

RALPH_INJECTIONS

RALPH_INJECTIONS_CRATES_IO

RALPH_LOCALS

RALPH_LOCALS_CRATES_IO

RASI_HIGHLIGHTS

RASI_HIGHLIGHTS_CRATES_IO

RASI_INJECTIONS

RASI_INJECTIONS_CRATES_IO

RASI_LOCALS

RASI_LOCALS_CRATES_IO

RAZOR_HIGHLIGHTS

RAZOR_HIGHLIGHTS_CRATES_IO

RAZOR_INJECTIONS

RAZOR_INJECTIONS_CRATES_IO

RAZOR_LOCALS

RAZOR_LOCALS_CRATES_IO

RBS_HIGHLIGHTS

RBS_HIGHLIGHTS_CRATES_IO

RBS_INJECTIONS

RBS_INJECTIONS_CRATES_IO

RBS_LOCALS

RBS_LOCALS_CRATES_IO

RE2C_HIGHLIGHTS

RE2C_HIGHLIGHTS_CRATES_IO

RE2C_INJECTIONS

RE2C_INJECTIONS_CRATES_IO

RE2C_LOCALS

RE2C_LOCALS_CRATES_IO

READLINE_HIGHLIGHTS

READLINE_HIGHLIGHTS_CRATES_IO

READLINE_INJECTIONS

READLINE_INJECTIONS_CRATES_IO

READLINE_LOCALS

READLINE_LOCALS_CRATES_IO

REGEX_HIGHLIGHTS

REGEX_HIGHLIGHTS_CRATES_IO

REGEX_INJECTIONS

REGEX_INJECTIONS_CRATES_IO

REGEX_LOCALS

REGEX_LOCALS_CRATES_IO

REGO_HIGHLIGHTS

REGO_HIGHLIGHTS_CRATES_IO

REGO_INJECTIONS

REGO_INJECTIONS_CRATES_IO

REGO_LOCALS

REGO_LOCALS_CRATES_IO

REQUIREMENTS_HIGHLIGHTS

REQUIREMENTS_HIGHLIGHTS_CRATES_IO

REQUIREMENTS_INJECTIONS

REQUIREMENTS_INJECTIONS_CRATES_IO

REQUIREMENTS_LOCALS

REQUIREMENTS_LOCALS_CRATES_IO

RESCRIPT_HIGHLIGHTS

RESCRIPT_HIGHLIGHTS_CRATES_IO

RESCRIPT_INJECTIONS

RESCRIPT_INJECTIONS_CRATES_IO

RESCRIPT_LOCALS

RESCRIPT_LOCALS_CRATES_IO

RIFLECONF_HIGHLIGHTS

RIFLECONF_HIGHLIGHTS_CRATES_IO

RIFLECONF_INJECTIONS

RIFLECONF_INJECTIONS_CRATES_IO

RIFLECONF_LOCALS

RIFLECONF_LOCALS_CRATES_IO

RNOWEB_HIGHLIGHTS

RNOWEB_HIGHLIGHTS_CRATES_IO

RNOWEB_INJECTIONS

RNOWEB_INJECTIONS_CRATES_IO

RNOWEB_LOCALS

RNOWEB_LOCALS_CRATES_IO

ROBOTS_HIGHLIGHTS

ROBOTS_HIGHLIGHTS_CRATES_IO

ROBOTS_INJECTIONS

ROBOTS_INJECTIONS_CRATES_IO

ROBOTS_LOCALS

ROBOTS_LOCALS_CRATES_IO

ROBOT_HIGHLIGHTS

ROBOT_HIGHLIGHTS_CRATES_IO

ROBOT_INJECTIONS

ROBOT_INJECTIONS_CRATES_IO

ROBOT_LOCALS

ROBOT_LOCALS_CRATES_IO

ROC_HIGHLIGHTS

ROC_HIGHLIGHTS_CRATES_IO

ROC_INJECTIONS

ROC_INJECTIONS_CRATES_IO

ROC_LOCALS

ROC_LOCALS_CRATES_IO

RON_HIGHLIGHTS

RON_HIGHLIGHTS_CRATES_IO

RON_INJECTIONS

RON_INJECTIONS_CRATES_IO

RON_LOCALS

RON_LOCALS_CRATES_IO

RST_HIGHLIGHTS

RST_HIGHLIGHTS_CRATES_IO

RST_INJECTIONS

RST_INJECTIONS_CRATES_IO

RST_LOCALS

RST_LOCALS_CRATES_IO

RUBY_HIGHLIGHTS

RUBY_HIGHLIGHTS_CRATES_IO

RUBY_INJECTIONS

RUBY_INJECTIONS_CRATES_IO

RUBY_LOCALS

RUBY_LOCALS_CRATES_IO

RUNESCRIPT_HIGHLIGHTS

RUNESCRIPT_HIGHLIGHTS_CRATES_IO

RUNESCRIPT_INJECTIONS

RUNESCRIPT_INJECTIONS_CRATES_IO

RUNESCRIPT_LOCALS

RUNESCRIPT_LOCALS_CRATES_IO

RUSH_HIGHLIGHTS

RUSH_HIGHLIGHTS_CRATES_IO

RUSH_INJECTIONS

RUSH_INJECTIONS_CRATES_IO

RUSH_LOCALS

RUSH_LOCALS_CRATES_IO

RUST_HIGHLIGHTS

RUST_HIGHLIGHTS_CRATES_IO

RUST_INJECTIONS

RUST_INJECTIONS_CRATES_IO

RUST_LOCALS

RUST_LOCALS_CRATES_IO

R_HIGHLIGHTS

R_HIGHLIGHTS_CRATES_IO

R_INJECTIONS

R_INJECTIONS_CRATES_IO

R_LOCALS

R_LOCALS_CRATES_IO

SCALA_HIGHLIGHTS

SCALA_HIGHLIGHTS_CRATES_IO

SCALA_INJECTIONS

SCALA_INJECTIONS_CRATES_IO

SCALA_LOCALS

SCALA_LOCALS_CRATES_IO

SCFG_HIGHLIGHTS

SCFG_HIGHLIGHTS_CRATES_IO

SCFG_INJECTIONS

SCFG_INJECTIONS_CRATES_IO

SCFG_LOCALS

SCFG_LOCALS_CRATES_IO

SCHEME_HIGHLIGHTS

SCHEME_HIGHLIGHTS_CRATES_IO

SCHEME_INJECTIONS

SCHEME_INJECTIONS_CRATES_IO

SCHEME_LOCALS

SCHEME_LOCALS_CRATES_IO

SCSS_HIGHLIGHTS

SCSS_HIGHLIGHTS_CRATES_IO

SCSS_INJECTIONS

SCSS_INJECTIONS_CRATES_IO

SCSS_LOCALS

SCSS_LOCALS_CRATES_IO

SFACE_HIGHLIGHTS

SFACE_HIGHLIGHTS_CRATES_IO

SFACE_INJECTIONS

SFACE_INJECTIONS_CRATES_IO

SFACE_LOCALS

SFACE_LOCALS_CRATES_IO

SFLOG_HIGHLIGHTS

SFLOG_HIGHLIGHTS_CRATES_IO

SFLOG_INJECTIONS

SFLOG_INJECTIONS_CRATES_IO

SFLOG_LOCALS

SFLOG_LOCALS_CRATES_IO

SHADERSLANG_HIGHLIGHTS

SHADERSLANG_HIGHLIGHTS_CRATES_IO

SHADERSLANG_INJECTIONS

SHADERSLANG_INJECTIONS_CRATES_IO

SHADERSLANG_LOCALS

SHADERSLANG_LOCALS_CRATES_IO

SLIM_HIGHLIGHTS

SLIM_HIGHLIGHTS_CRATES_IO

SLIM_INJECTIONS

SLIM_INJECTIONS_CRATES_IO

SLIM_LOCALS

SLIM_LOCALS_CRATES_IO

SLINT_HIGHLIGHTS

SLINT_HIGHLIGHTS_CRATES_IO

SLINT_INJECTIONS

SLINT_INJECTIONS_CRATES_IO

SLINT_LOCALS

SLINT_LOCALS_CRATES_IO

SMALI_HIGHLIGHTS

SMALI_HIGHLIGHTS_CRATES_IO

SMALI_INJECTIONS

SMALI_INJECTIONS_CRATES_IO

SMALI_LOCALS

SMALI_LOCALS_CRATES_IO

SMITHY_HIGHLIGHTS

SMITHY_HIGHLIGHTS_CRATES_IO

SMITHY_INJECTIONS

SMITHY_INJECTIONS_CRATES_IO

SMITHY_LOCALS

SMITHY_LOCALS_CRATES_IO

SNAKEMAKE_HIGHLIGHTS

SNAKEMAKE_HIGHLIGHTS_CRATES_IO

SNAKEMAKE_INJECTIONS

SNAKEMAKE_INJECTIONS_CRATES_IO

SNAKEMAKE_LOCALS

SNAKEMAKE_LOCALS_CRATES_IO

SOLIDITY_HIGHLIGHTS

SOLIDITY_HIGHLIGHTS_CRATES_IO

SOLIDITY_INJECTIONS

SOLIDITY_INJECTIONS_CRATES_IO

SOLIDITY_LOCALS

SOLIDITY_LOCALS_CRATES_IO

SOQL_HIGHLIGHTS

SOQL_HIGHLIGHTS_CRATES_IO

SOQL_INJECTIONS

SOQL_INJECTIONS_CRATES_IO

SOQL_LOCALS

SOQL_LOCALS_CRATES_IO

SOSL_HIGHLIGHTS

SOSL_HIGHLIGHTS_CRATES_IO

SOSL_INJECTIONS

SOSL_INJECTIONS_CRATES_IO

SOSL_LOCALS

SOSL_LOCALS_CRATES_IO

SOURCEPAWN_HIGHLIGHTS

SOURCEPAWN_HIGHLIGHTS_CRATES_IO

SOURCEPAWN_INJECTIONS

SOURCEPAWN_INJECTIONS_CRATES_IO

SOURCEPAWN_LOCALS

SOURCEPAWN_LOCALS_CRATES_IO

SPARQL_HIGHLIGHTS

SPARQL_HIGHLIGHTS_CRATES_IO

SPARQL_INJECTIONS

SPARQL_INJECTIONS_CRATES_IO

SPARQL_LOCALS

SPARQL_LOCALS_CRATES_IO

SPROTO_HIGHLIGHTS

SPROTO_HIGHLIGHTS_CRATES_IO

SPROTO_INJECTIONS

SPROTO_INJECTIONS_CRATES_IO

SPROTO_LOCALS

SPROTO_LOCALS_CRATES_IO

SQL_HIGHLIGHTS

SQL_HIGHLIGHTS_CRATES_IO

SQL_INJECTIONS

SQL_INJECTIONS_CRATES_IO

SQL_LOCALS

SQL_LOCALS_CRATES_IO

SQUIRREL_HIGHLIGHTS

SQUIRREL_HIGHLIGHTS_CRATES_IO

SQUIRREL_INJECTIONS

SQUIRREL_INJECTIONS_CRATES_IO

SQUIRREL_LOCALS

SQUIRREL_LOCALS_CRATES_IO

SSHCONFIG_HIGHLIGHTS

SSHCONFIG_HIGHLIGHTS_CRATES_IO

SSHCONFIG_INJECTIONS

SSHCONFIG_INJECTIONS_CRATES_IO

SSHCONFIG_LOCALS

SSHCONFIG_LOCALS_CRATES_IO

STRACE_HIGHLIGHTS

STRACE_HIGHLIGHTS_CRATES_IO

STRACE_INJECTIONS

STRACE_INJECTIONS_CRATES_IO

STRACE_LOCALS

STRACE_LOCALS_CRATES_IO

STYLED_HIGHLIGHTS

STYLED_HIGHLIGHTS_CRATES_IO

STYLED_INJECTIONS

STYLED_INJECTIONS_CRATES_IO

STYLED_LOCALS

STYLED_LOCALS_CRATES_IO

SUPERCOLLIDER_HIGHLIGHTS

SUPERCOLLIDER_HIGHLIGHTS_CRATES_IO

SUPERCOLLIDER_INJECTIONS

SUPERCOLLIDER_INJECTIONS_CRATES_IO

SUPERCOLLIDER_LOCALS

SUPERCOLLIDER_LOCALS_CRATES_IO

SUPERHTML_HIGHLIGHTS

SUPERHTML_HIGHLIGHTS_CRATES_IO

SUPERHTML_INJECTIONS

SUPERHTML_INJECTIONS_CRATES_IO

SUPERHTML_LOCALS

SUPERHTML_LOCALS_CRATES_IO

SVELTE_HIGHLIGHTS

SVELTE_HIGHLIGHTS_CRATES_IO

SVELTE_INJECTIONS

SVELTE_INJECTIONS_CRATES_IO

SVELTE_LOCALS

SVELTE_LOCALS_CRATES_IO

SWAY_HIGHLIGHTS

SWAY_HIGHLIGHTS_CRATES_IO

SWAY_INJECTIONS

SWAY_INJECTIONS_CRATES_IO

SWAY_LOCALS

SWAY_LOCALS_CRATES_IO

SWIFT_HIGHLIGHTS

SWIFT_HIGHLIGHTS_CRATES_IO

SWIFT_INJECTIONS

SWIFT_INJECTIONS_CRATES_IO

SWIFT_LOCALS

SWIFT_LOCALS_CRATES_IO

SXHKDRC_HIGHLIGHTS

SXHKDRC_HIGHLIGHTS_CRATES_IO

SXHKDRC_INJECTIONS

SXHKDRC_INJECTIONS_CRATES_IO

SXHKDRC_LOCALS

SXHKDRC_LOCALS_CRATES_IO

SYSTEMTAP_HIGHLIGHTS

SYSTEMTAP_HIGHLIGHTS_CRATES_IO

SYSTEMTAP_INJECTIONS

SYSTEMTAP_INJECTIONS_CRATES_IO

SYSTEMTAP_LOCALS

SYSTEMTAP_LOCALS_CRATES_IO

TABLEGEN_HIGHLIGHTS

TABLEGEN_HIGHLIGHTS_CRATES_IO

TABLEGEN_INJECTIONS

TABLEGEN_INJECTIONS_CRATES_IO

TABLEGEN_LOCALS

TABLEGEN_LOCALS_CRATES_IO

TACT_HIGHLIGHTS

TACT_HIGHLIGHTS_CRATES_IO

TACT_INJECTIONS

TACT_INJECTIONS_CRATES_IO

TACT_LOCALS

TACT_LOCALS_CRATES_IO

TAL_HIGHLIGHTS

TAL_HIGHLIGHTS_CRATES_IO

TAL_INJECTIONS

TAL_INJECTIONS_CRATES_IO

TAL_LOCALS

TAL_LOCALS_CRATES_IO

TAPE_HIGHLIGHTS

TAPE_HIGHLIGHTS_CRATES_IO

TAPE_INJECTIONS

TAPE_INJECTIONS_CRATES_IO

TAPE_LOCALS

TAPE_LOCALS_CRATES_IO

TCL_HIGHLIGHTS

TCL_HIGHLIGHTS_CRATES_IO

TCL_INJECTIONS

TCL_INJECTIONS_CRATES_IO

TCL_LOCALS

TCL_LOCALS_CRATES_IO

TEAL_HIGHLIGHTS

TEAL_HIGHLIGHTS_CRATES_IO

TEAL_INJECTIONS

TEAL_INJECTIONS_CRATES_IO

TEAL_LOCALS

TEAL_LOCALS_CRATES_IO

TEMPL_HIGHLIGHTS

TEMPL_HIGHLIGHTS_CRATES_IO

TEMPL_INJECTIONS

TEMPL_INJECTIONS_CRATES_IO

TEMPL_LOCALS

TEMPL_LOCALS_CRATES_IO

TERA_HIGHLIGHTS

TERA_HIGHLIGHTS_CRATES_IO

TERA_INJECTIONS

TERA_INJECTIONS_CRATES_IO

TERA_LOCALS

TERA_LOCALS_CRATES_IO

TERRAFORM_HIGHLIGHTS

TERRAFORM_HIGHLIGHTS_CRATES_IO

TERRAFORM_INJECTIONS

TERRAFORM_INJECTIONS_CRATES_IO

TERRAFORM_LOCALS

TERRAFORM_LOCALS_CRATES_IO

TEX_HIGHLIGHTS

TEX_HIGHLIGHTS_CRATES_IO

TEX_INJECTIONS

TEX_INJECTIONS_CRATES_IO

TEX_LOCALS

TEX_LOCALS_CRATES_IO

THRIFT_HIGHLIGHTS

THRIFT_HIGHLIGHTS_CRATES_IO

THRIFT_INJECTIONS

THRIFT_INJECTIONS_CRATES_IO

THRIFT_LOCALS

THRIFT_LOCALS_CRATES_IO

TIGER_HIGHLIGHTS

TIGER_HIGHLIGHTS_CRATES_IO

TIGER_INJECTIONS

TIGER_INJECTIONS_CRATES_IO

TIGER_LOCALS

TIGER_LOCALS_CRATES_IO

TLA_HIGHLIGHTS

TLA_HIGHLIGHTS_CRATES_IO

TLA_INJECTIONS

TLA_INJECTIONS_CRATES_IO

TLA_LOCALS

TLA_LOCALS_CRATES_IO

TMUX_HIGHLIGHTS

TMUX_HIGHLIGHTS_CRATES_IO

TMUX_INJECTIONS

TMUX_INJECTIONS_CRATES_IO

TMUX_LOCALS

TMUX_LOCALS_CRATES_IO

TODOTXT_HIGHLIGHTS

TODOTXT_HIGHLIGHTS_CRATES_IO

TODOTXT_INJECTIONS

TODOTXT_INJECTIONS_CRATES_IO

TODOTXT_LOCALS

TODOTXT_LOCALS_CRATES_IO

TOML_HIGHLIGHTS

TOML_HIGHLIGHTS_CRATES_IO

TOML_INJECTIONS

TOML_INJECTIONS_CRATES_IO

TOML_LOCALS

TOML_LOCALS_CRATES_IO

TRACE32_HIGHLIGHTS

TRACE32_HIGHLIGHTS_CRATES_IO

TRACE32_INJECTIONS

TRACE32_INJECTIONS_CRATES_IO

TRACE32_LOCALS

TRACE32_LOCALS_CRATES_IO

TSV_HIGHLIGHTS

TSV_HIGHLIGHTS_CRATES_IO

TSV_INJECTIONS

TSV_INJECTIONS_CRATES_IO

TSV_LOCALS

TSV_LOCALS_CRATES_IO

TSX_HIGHLIGHTS

TSX_HIGHLIGHTS_CRATES_IO

TSX_INJECTIONS

TSX_INJECTIONS_CRATES_IO

TSX_LOCALS

TSX_LOCALS_CRATES_IO

TURTLE_HIGHLIGHTS

TURTLE_HIGHLIGHTS_CRATES_IO

TURTLE_INJECTIONS

TURTLE_INJECTIONS_CRATES_IO

TURTLE_LOCALS

TURTLE_LOCALS_CRATES_IO

TWIG_HIGHLIGHTS

TWIG_HIGHLIGHTS_CRATES_IO

TWIG_INJECTIONS

TWIG_INJECTIONS_CRATES_IO

TWIG_LOCALS

TWIG_LOCALS_CRATES_IO

TYPESCRIPT_HIGHLIGHTS

TYPESCRIPT_HIGHLIGHTS_CRATES_IO

TYPESCRIPT_INJECTIONS

TYPESCRIPT_INJECTIONS_CRATES_IO

TYPESCRIPT_LOCALS

TYPESCRIPT_LOCALS_CRATES_IO

TYPESPEC_HIGHLIGHTS

TYPESPEC_HIGHLIGHTS_CRATES_IO

TYPESPEC_INJECTIONS

TYPESPEC_INJECTIONS_CRATES_IO

TYPESPEC_LOCALS

TYPESPEC_LOCALS_CRATES_IO

TYPOSCRIPT_HIGHLIGHTS

TYPOSCRIPT_HIGHLIGHTS_CRATES_IO

TYPOSCRIPT_INJECTIONS

TYPOSCRIPT_INJECTIONS_CRATES_IO

TYPOSCRIPT_LOCALS

TYPOSCRIPT_LOCALS_CRATES_IO

TYPST_HIGHLIGHTS

TYPST_HIGHLIGHTS_CRATES_IO

TYPST_INJECTIONS

TYPST_INJECTIONS_CRATES_IO

TYPST_LOCALS

TYPST_LOCALS_CRATES_IO

UDEVRULES_HIGHLIGHTS

UDEVRULES_HIGHLIGHTS_CRATES_IO

UDEVRULES_INJECTIONS

UDEVRULES_INJECTIONS_CRATES_IO

UDEVRULES_LOCALS

UDEVRULES_LOCALS_CRATES_IO

UNGRAMMAR_HIGHLIGHTS

UNGRAMMAR_HIGHLIGHTS_CRATES_IO

UNGRAMMAR_INJECTIONS

UNGRAMMAR_INJECTIONS_CRATES_IO

UNGRAMMAR_LOCALS

UNGRAMMAR_LOCALS_CRATES_IO

UNISON_HIGHLIGHTS

UNISON_HIGHLIGHTS_CRATES_IO

UNISON_INJECTIONS

UNISON_INJECTIONS_CRATES_IO

UNISON_LOCALS

UNISON_LOCALS_CRATES_IO

URSA_HIGHLIGHTS

URSA_HIGHLIGHTS_CRATES_IO

URSA_INJECTIONS

URSA_INJECTIONS_CRATES_IO

URSA_LOCALS

URSA_LOCALS_CRATES_IO

USD_HIGHLIGHTS

USD_HIGHLIGHTS_CRATES_IO

USD_INJECTIONS

USD_INJECTIONS_CRATES_IO

USD_LOCALS

USD_LOCALS_CRATES_IO

VALA_HIGHLIGHTS

VALA_HIGHLIGHTS_CRATES_IO

VALA_INJECTIONS

VALA_INJECTIONS_CRATES_IO

VALA_LOCALS

VALA_LOCALS_CRATES_IO

VERILOG_HIGHLIGHTS

VERILOG_HIGHLIGHTS_CRATES_IO

VERILOG_INJECTIONS

VERILOG_INJECTIONS_CRATES_IO

VERILOG_LOCALS

VERILOG_LOCALS_CRATES_IO

VHDL_HIGHLIGHTS

VHDL_HIGHLIGHTS_CRATES_IO

VHDL_INJECTIONS

VHDL_INJECTIONS_CRATES_IO

VHDL_LOCALS

VHDL_LOCALS_CRATES_IO

VIM_HIGHLIGHTS

VIM_HIGHLIGHTS_CRATES_IO

VIM_INJECTIONS

VIM_INJECTIONS_CRATES_IO

VIM_LOCALS

VIM_LOCALS_CRATES_IO

VRL_HIGHLIGHTS

VRL_HIGHLIGHTS_CRATES_IO

VRL_INJECTIONS

VRL_INJECTIONS_CRATES_IO

VRL_LOCALS

VRL_LOCALS_CRATES_IO

VTO_HIGHLIGHTS

VTO_HIGHLIGHTS_CRATES_IO

VTO_INJECTIONS

VTO_INJECTIONS_CRATES_IO

VTO_LOCALS

VTO_LOCALS_CRATES_IO

VUE_HIGHLIGHTS

VUE_HIGHLIGHTS_CRATES_IO

VUE_INJECTIONS

VUE_INJECTIONS_CRATES_IO

VUE_LOCALS

VUE_LOCALS_CRATES_IO

V_HIGHLIGHTS

V_HIGHLIGHTS_CRATES_IO

V_INJECTIONS

V_INJECTIONS_CRATES_IO

V_LOCALS

V_LOCALS_CRATES_IO

WAT_HIGHLIGHTS

WAT_HIGHLIGHTS_CRATES_IO

WAT_INJECTIONS

WAT_INJECTIONS_CRATES_IO

WAT_LOCALS

WAT_LOCALS_CRATES_IO

WGSL_BEVY_HIGHLIGHTS

WGSL_BEVY_HIGHLIGHTS_CRATES_IO

WGSL_BEVY_INJECTIONS

WGSL_BEVY_INJECTIONS_CRATES_IO

WGSL_BEVY_LOCALS

WGSL_BEVY_LOCALS_CRATES_IO

WGSL_HIGHLIGHTS

WGSL_HIGHLIGHTS_CRATES_IO

WGSL_INJECTIONS

WGSL_INJECTIONS_CRATES_IO

WGSL_LOCALS

WGSL_LOCALS_CRATES_IO

WING_HIGHLIGHTS

WING_HIGHLIGHTS_CRATES_IO

WING_INJECTIONS

WING_INJECTIONS_CRATES_IO

WING_LOCALS

WING_LOCALS_CRATES_IO

WIT_HIGHLIGHTS

WIT_HIGHLIGHTS_CRATES_IO

WIT_INJECTIONS

WIT_INJECTIONS_CRATES_IO

WIT_LOCALS

WIT_LOCALS_CRATES_IO

WXML_HIGHLIGHTS

WXML_HIGHLIGHTS_CRATES_IO

WXML_INJECTIONS

WXML_INJECTIONS_CRATES_IO

WXML_LOCALS

WXML_LOCALS_CRATES_IO

XCOMPOSE_HIGHLIGHTS

XCOMPOSE_HIGHLIGHTS_CRATES_IO

XCOMPOSE_INJECTIONS

XCOMPOSE_INJECTIONS_CRATES_IO

XCOMPOSE_LOCALS

XCOMPOSE_LOCALS_CRATES_IO

XDEFAULTS_HIGHLIGHTS

XDEFAULTS_HIGHLIGHTS_CRATES_IO

XDEFAULTS_INJECTIONS

XDEFAULTS_INJECTIONS_CRATES_IO

XDEFAULTS_LOCALS

XDEFAULTS_LOCALS_CRATES_IO

XML_HIGHLIGHTS

XML_HIGHLIGHTS_CRATES_IO

XML_INJECTIONS

XML_INJECTIONS_CRATES_IO

XML_LOCALS

XML_LOCALS_CRATES_IO

YAML_HIGHLIGHTS

YAML_HIGHLIGHTS_CRATES_IO

YAML_INJECTIONS

YAML_INJECTIONS_CRATES_IO

YAML_LOCALS

YAML_LOCALS_CRATES_IO

YANG_HIGHLIGHTS

YANG_HIGHLIGHTS_CRATES_IO

YANG_INJECTIONS

YANG_INJECTIONS_CRATES_IO

YANG_LOCALS

YANG_LOCALS_CRATES_IO

YUCK_HIGHLIGHTS

YUCK_HIGHLIGHTS_CRATES_IO

YUCK_INJECTIONS

YUCK_INJECTIONS_CRATES_IO

YUCK_LOCALS

YUCK_LOCALS_CRATES_IO

ZATHURARC_HIGHLIGHTS

ZATHURARC_HIGHLIGHTS_CRATES_IO

ZATHURARC_INJECTIONS

ZATHURARC_INJECTIONS_CRATES_IO

ZATHURARC_LOCALS

ZATHURARC_LOCALS_CRATES_IO

ZIGGY_HIGHLIGHTS

ZIGGY_HIGHLIGHTS_CRATES_IO

ZIGGY_INJECTIONS

ZIGGY_INJECTIONS_CRATES_IO

ZIGGY_LOCALS

ZIGGY_LOCALS_CRATES_IO

ZIGGY_SCHEMA_HIGHLIGHTS

ZIGGY_SCHEMA_HIGHLIGHTS_CRATES_IO

ZIGGY_SCHEMA_INJECTIONS

ZIGGY_SCHEMA_INJECTIONS_CRATES_IO

ZIGGY_SCHEMA_LOCALS

ZIGGY_SCHEMA_LOCALS_CRATES_IO

ZIG_HIGHLIGHTS

ZIG_HIGHLIGHTS_CRATES_IO

ZIG_INJECTIONS

ZIG_INJECTIONS_CRATES_IO

ZIG_LOCALS

ZIG_LOCALS_CRATES_IO

ZSH_HIGHLIGHTS

ZSH_HIGHLIGHTS_CRATES_IO

ZSH_INJECTIONS

ZSH_INJECTIONS_CRATES_IO

ZSH_LOCALS

ZSH_LOCALS_CRATES_IO